code redesign
